+++ CHANGELOG.rst
@@ -3,6 +3,7 @@ Release 0.4 UNRELEASED
 * fixed some packaging bugs.
 * made the data model more tolerant to missing symbols.
 * fixed symbol resolving bug.
+* added output of unused symbols.
 
 Release 0.3 16/04/2013
 -----------------------